Correccion validaciones, revocación de tokens, url mongo

+++ inventarios_backend/controllers/estadoEquipoControllers.js
@@ -1,4 +1,5 @@
 import EstadoEquipo from "../models/EstadoEquipo.js";
+import Inventario from "../models/Inventario.js";
 
 export const listarEstadosEquipo = async () => {
   return await EstadoEquipo.find();
@@ -16,6 +17,8 @@ export const editarEstadoEquipo = async (estadoEquipoId, estadoEquipoRequest) =>
 }
 
 export const eliminarEstadoEquipo = async (id) => {
+  const conteoReferencias = await Inventario.countDocuments({estadoEquipo: id});
+  if(conteoReferencias > 0) return {deleted: false, referencias: conteoReferencias}
   let estadoEquipoEliminado = await EstadoEquipo.findByIdAndDelete(id);
   return (estadoEquipoEliminado !== null) ? {deleted: true} : {deleted: false};
 }
