+++ inventarios_backend/controllers/tipoEquipoControllers.js
@@ -1,3 +1,4 @@
+import Inventario from "../models/Inventario.js";
 import TipoEquipo from "../models/TipoEquipo.js";
 
 export const listarTiposEquipo = async () => {
@@ -16,6 +17,8 @@ export const editarTipoEquipo = async (tipoEquipoId, tipoEquipoRequest) => {
 }
 
 export const eliminarTipoEquipo = async (id) => {
+  const conteoReferencias = await Inventario.countDocuments({tipoEquipo: id});
+  if(conteoReferencias > 0) return {deleted: false, referencias: conteoReferencias}
   let tipoEquipoEliminado = await TipoEquipo.findByIdAndDelete(id);
   return (tipoEquipoEliminado !== null) ? {deleted: true} : {deleted: false};
 }
