+++ inventarios_backend/controllers/usuarioControllers.js
@@ -1,6 +1,7 @@
 import Usuario from "../models/Usuario.js";
 import bcrypt from 'bcrypt';
 import UsuarioToken from "../models/UsuarioToken.js";
+import Inventario from "../models/Inventario.js";
 
 export const listarUsuarios = async () => {
   return await Usuario.find();
@@ -18,10 +19,13 @@ export const crearUsuario = async (usuarioRequest) => {
 
 export const editarUsuario = async (usuarioId, usuarioRequest) => {
   usuarioRequest.fechaActualizacion = new Date();
-  await Usuario.findByIdAndUpdate(usuarioId, usuarioRequest);
+  const usuario = await Usuario.findByIdAndUpdate(usuarioId, usuarioRequest, {new: true});
+  return {_id: usuario._id, rol: usuario.rol}
 }
 
 export const eliminarUsuario = async (id) => {
+  const conteoReferencias = await Inventario.countDocuments({usuario: id});
+  if(conteoReferencias > 0) return {deleted: false, referencias: conteoReferencias}
   let usuarioEliminado = await Usuario.findByIdAndDelete(id);
   if(usuarioEliminado) {
     await UsuarioToken.findOneAndDelete( {usuarioId: id} );
